--- packages/meditrak-server/src/tests/routeTests/index.js
@@ -1,8 +1,0 @@
-/**
- * Tupaia MediTrak
- * Copyright (c) 2017 Beyond Essential Systems Pty Ltd
- */
-
-export { testGetChangesCount } from './changes/get';
-export { testPostChanges } from './changes/post';
-export { testPutSurveyResponses } from './surveyResponse/put';
