+++ packages/meditrak-server/src/tests/routes/GETAnswers.test.js
@@ -9,10 +9,8 @@ import {
   buildAndInsertSurveyResponses,
   findOrCreateDummyRecord,
 } from '@tupaia/database';
-import { Authenticator } from '@tupaia/auth';
-import { resetTestData } from '../testUtilities';
+import { resetTestData, TestableApp } from '../testUtilities';
 import { TUPAIA_ADMIN_PANEL_PERMISSION_GROUP, BES_ADMIN_PERMISSION_GROUP } from '../../permissions';
-import { TestableApp } from '../TestableApp';
 
 describe('Permissions checker for GETAnswers', async () => {
   const DEFAULT_POLICY = {
