+++ packages/meditrak-server/src/tests/routes/GETDashboardGroups.test.js
@@ -9,9 +9,8 @@ import {
   buildAndInsertProjectsAndHierarchies,
   addBaselineTestCountries,
 } from '@tupaia/database';
-import { Authenticator } from '@tupaia/auth';
 import { TUPAIA_ADMIN_PANEL_PERMISSION_GROUP, BES_ADMIN_PERMISSION_GROUP } from '../../permissions';
-import { TestableApp } from '../TestableApp';
+import { TestableApp } from '../testUtilities';
 
 describe('Permissions checker for GETDashboardGroups', async () => {
   const DEFAULT_POLICY = {
