+++ packages/meditrak-server/src/tests/routes/GETDashboardReports.test.js
@@ -9,9 +9,8 @@ import {
   addBaselineTestCountries,
   buildAndInsertProjectsAndHierarchies,
 } from '@tupaia/database';
-import { Authenticator } from '@tupaia/auth';
 import { TUPAIA_ADMIN_PANEL_PERMISSION_GROUP, BES_ADMIN_PERMISSION_GROUP } from '../../permissions';
-import { TestableApp } from '../TestableApp';
+import { TestableApp } from '../testUtilities';
 
 describe('Permissions checker for GETDashboardReports', async () => {
   const DEFAULT_POLICY = {
