+++ packages/meditrak-server/src/tests/routes/GETDisasters.test.js
@@ -6,7 +6,7 @@
 import { expect } from 'chai';
 import { findOrCreateDummyRecord } from '@tupaia/database';
 import { TUPAIA_ADMIN_PANEL_PERMISSION_GROUP, BES_ADMIN_PERMISSION_GROUP } from '../../permissions';
-import { TestableApp } from '../TestableApp';
+import { TestableApp } from '../testUtilities';
 
 const getFilterString = filter => `filter=${JSON.stringify(filter)}`;
