+++ packages/meditrak-server/src/tests/routes/GETMapOverlays.test.js
@@ -10,7 +10,7 @@ import {
   buildAndInsertProjectsAndHierarchies,
 } from '@tupaia/database';
 import { TUPAIA_ADMIN_PANEL_PERMISSION_GROUP, BES_ADMIN_PERMISSION_GROUP } from '../../permissions';
-import { TestableApp } from '../TestableApp';
+import { TestableApp } from '../testUtilities';
 
 describe('Permissions checker for GETMapOverlays', async () => {
   const DEFAULT_POLICY = {
