+++ packages/meditrak-server/src/tests/routes/GETSurveys.test.js
@@ -5,9 +5,8 @@
 
 import { expect } from 'chai';
 import { buildAndInsertSurveys, findOrCreateDummyRecord } from '@tupaia/database';
-import { resetTestData } from '../testUtilities';
+import { resetTestData, TestableApp } from '../testUtilities';
 import { TUPAIA_ADMIN_PANEL_PERMISSION_GROUP, BES_ADMIN_PERMISSION_GROUP } from '../../permissions';
-import { TestableApp } from '../TestableApp';
 
 describe('Permissions checker for GETSurveys', async () => {
   const DEFAULT_POLICY = {
