+++ packages/meditrak-server/src/tests/routes/authenticate/oneTimeLogin.test.js
@@ -7,8 +7,7 @@ import {} from 'dotenv/config'; // Load the environment variables into process.e
 import { expect } from 'chai';
 import moment from 'moment';
 
-import { TestableApp, getAuthorizationHeader } from './TestableApp';
-import { randomEmail } from './testUtilities';
+import { getAuthorizationHeader, randomEmail, TestableApp } from '../../testUtilities';
 
 describe('One Time Login', function () {
   const app = new TestableApp();
