+++ packages/meditrak-server/src/tests/routes/exportSurveys/exportSurveys.test.js
@@ -6,14 +6,13 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 import xlsx from 'xlsx';
-import { Authenticator } from '@tupaia/auth';
 import {
   buildAndInsertSurveys,
   findOrCreateDummyRecord,
   findOrCreateDummyCountryEntity,
 } from '@tupaia/database';
 import { TUPAIA_ADMIN_PANEL_PERMISSION_GROUP } from '../../../permissions';
-import { TestableApp } from '../../TestableApp';
+import { TestableApp } from '../../testUtilities';
 
 const DEFAULT_POLICY = {
   DL: ['Public'],
