+++ packages/meditrak-server/src/tests/routes/importSurveys/assertCanImportSurveys.test.js
@@ -11,7 +11,7 @@ import {
   findOrCreateDummyCountryEntity,
 } from '@tupaia/database';
 import { TUPAIA_ADMIN_PANEL_PERMISSION_GROUP } from '../../../permissions';
-import { getModels } from '../../getModels';
+import { getModels } from '../../testUtilities';
 import { assertCanImportSurveys } from '../../../routes/importSurveys/assertCanImportSurveys';
 
 const DEFAULT_POLICY = {
@@ -60,14 +60,14 @@ describe('assertCanImportSurveys(): Permissions checker for Importing Surveys', 
     const accessPolicy = new AccessPolicy(DEFAULT_POLICY);
 
     it('Should allow importing an existing survey if users have Tupaia Admin Panel and survey permission group access to the country of that survey', async () => {
-      const surveyNames = [SURVEY_NAME_1]; //Existing survey, already inserted
+      const surveyNames = [SURVEY_NAME_1]; // Existing survey, already inserted
       const result = await assertCanImportSurveys(accessPolicy, models, surveyNames);
 
       expect(result).to.true;
     });
 
     it('Should allow import an existing survey and update countryIds if users have Tupaia Admin Panel and survey permission group access to the country of that survey', async () => {
-      const surveyNames = [SURVEY_NAME_1]; //Existing survey, already inserted
+      const surveyNames = [SURVEY_NAME_1]; // Existing survey, already inserted
       const newCountryIds = [kiribatiCountry.id];
       const result = await assertCanImportSurveys(accessPolicy, models, surveyNames, newCountryIds);
 
@@ -93,7 +93,7 @@ describe('assertCanImportSurveys(): Permissions checker for Importing Surveys', 
         policyWithAllCountries[c.code] = [TUPAIA_ADMIN_PANEL_PERMISSION_GROUP];
       });
       const allCountriesAccessPolicy = new AccessPolicy(policyWithAllCountries);
-      const surveyNames = [SURVEY_NAME_3]; //New survey
+      const surveyNames = [SURVEY_NAME_3]; // New survey
       const result = await assertCanImportSurveys(allCountriesAccessPolicy, models, surveyNames);
 
       expect(result).to.true;
@@ -110,7 +110,7 @@ describe('assertCanImportSurveys(): Permissions checker for Importing Surveys', 
         LA: ['Admin'],
       };
       const accessPolicy = new AccessPolicy(policy);
-      const surveyNames = [SURVEY_NAME_1]; //Existing survey, already inserted
+      const surveyNames = [SURVEY_NAME_1]; // Existing survey, already inserted
 
       expect(() => assertCanImportSurveys(accessPolicy, models, surveyNames)).to.throw;
     });
@@ -124,7 +124,7 @@ describe('assertCanImportSurveys(): Permissions checker for Importing Surveys', 
         LA: ['Admin'],
       };
       const accessPolicy = new AccessPolicy(policy);
-      const surveyNames = [SURVEY_NAME_1]; //Existing survey, already
+      const surveyNames = [SURVEY_NAME_1]; // Existing survey, already
 
       expect(() => assertCanImportSurveys(accessPolicy, models, surveyNames)).to.throw;
     });
@@ -138,7 +138,7 @@ describe('assertCanImportSurveys(): Permissions checker for Importing Surveys', 
         LA: ['Admin'],
       };
       const accessPolicy = new AccessPolicy(policy);
-      const surveyNames = [SURVEY_NAME_1]; //Existing survey, already inserted
+      const surveyNames = [SURVEY_NAME_1]; // Existing survey, already inserted
       const newCountryIds = [kiribatiCountry.id];
 
       expect(() => assertCanImportSurveys(accessPolicy, models, surveyNames, newCountryIds)).to
@@ -156,7 +156,7 @@ describe('assertCanImportSurveys(): Permissions checker for Importing Surveys', 
         LA: ['Admin'],
       };
       const accessPolicy = new AccessPolicy(policy);
-      const surveyNames = [SURVEY_NAME_2]; //New survey
+      const surveyNames = [SURVEY_NAME_2]; // New survey
       const newCountryIds = [vanuatuCountry.id];
 
       expect(() => assertCanImportSurveys(accessPolicy, models, surveyNames, newCountryIds)).to
@@ -165,7 +165,7 @@ describe('assertCanImportSurveys(): Permissions checker for Importing Surveys', 
 
     it('Should not allow importing a new survey with no countries are specified when users have Tupaia Admin Panel access to all countries', async () => {
       const accessPolicy = new AccessPolicy(DEFAULT_POLICY);
-      const surveyNames = [SURVEY_NAME_3]; //New survey
+      const surveyNames = [SURVEY_NAME_3]; // New survey
 
       expect(() => assertCanImportSurveys(accessPolicy, models, surveyNames)).to.throw;
     });
