+++ packages/meditrak-server/src/tests/routes/importSurveys/importSurveys.test.js
@@ -4,19 +4,16 @@
  */
 
 import { expect } from 'chai';
-import sinon from 'sinon';
-import { Authenticator } from '@tupaia/auth';
 import {
   buildAndInsertSurveys,
   findOrCreateDummyRecord,
   findOrCreateDummyCountryEntity,
 } from '@tupaia/database';
-import { resetTestData } from '../../testUtilities';
+import { resetTestData, TestableApp } from '../../testUtilities';
 import {
   TUPAIA_ADMIN_PANEL_PERMISSION_GROUP,
   BES_ADMIN_PERMISSION_GROUP,
 } from '../../../permissions';
-import { TestableApp } from '../../TestableApp';
 import { expectPermissionError } from '../../testUtilities/expectResponseError';
 
 const DEFAULT_POLICY = {
