+++ packages/meditrak-server/src/tests/routes/userAccounts/GETUserAccounts.test.js
@@ -9,7 +9,7 @@ import {
   TUPAIA_ADMIN_PANEL_PERMISSION_GROUP,
   BES_ADMIN_PERMISSION_GROUP,
 } from '../../../permissions';
-import { TestableApp } from '../../TestableApp';
+import { TestableApp } from '../../testUtilities';
 
 describe('Permissions checker for GETUserAccounts', async () => {
   const DEFAULT_POLICY = {
