+++ packages/meditrak-server/src/tests/testUtilities/database/index.js
@@ -3,18 +3,7 @@
  * Copyright (c) 2019 Beyond Essential Systems Pty Ltd
  */
 
+export { getModels } from './getModels';
 export { insertEntityAndFacility } from './insertEntityAndFacility';
 export { resetTestData } from './resetTestData';
-export {
-  upsertAnswer,
-  upsertEntity,
-  upsertQuestion,
-  upsertDataSource,
-  upsertSurvey,
-  upsertSurveyResponse,
-  upsertSurveyScreen,
-  upsertSurveyScreenComponent,
-  upsertUserEntityPermission,
-  upsertComment,
-  upsertSurveyResponseComment,
-} from './upsertRecord';
+export * from './upsertRecord';
