+++ packages/meditrak-server/src/tests/testUtilities/database/resetTestData.js
@@ -4,7 +4,7 @@
  */
 import { clearTestData } from '@tupaia/database';
 import { addBaselineTestData } from './addBaselineTestData';
-import { getModels } from '../../getModels';
+import { getModels } from './getModels';
 
 export async function resetTestData() {
   const models = getModels();
