+++ packages/meditrak-server/src/tests/testUtilities/index.js
@@ -3,20 +3,8 @@
  * Copyright (c) 2019 Beyond Essential Systems Pty Ltd
  */
 
-export {
-  insertEntityAndFacility,
-  resetTestData,
-  upsertAnswer,
-  upsertEntity,
-  upsertQuestion,
-  upsertDataSource,
-  upsertSurvey,
-  upsertSurveyResponse,
-  upsertSurveyScreen,
-  upsertSurveyScreenComponent,
-  upsertComment,
-  upsertSurveyResponseComment,
-  upsertUserEntityPermission,
-} from './database';
+export * from './database';
+export { expectPermissionError, expectResponseError } from './expectResponseError';
 export { randomEmail, randomIntBetween, randomString } from './random';
 export { setupDummySyncQueue } from './setupDummySyncQueue';
+export { getAuthorizationHeader, TestableApp } from './TestableApp';
